refactor: separate documents and publishing capabilities
