+++ backend/services/documents/__init__.py
@@ -1,0 +1,19 @@
+"""Document parsing, knowledge, and book processing capabilities."""
+
+from .book_scanner_service import BookScannerService
+from .file_parser_service import FileParserService, get_file_parser, init_file_parser
+from .knowledge_service import (
+    KnowledgeService,
+    get_knowledge_service,
+    init_knowledge_service,
+)
+
+__all__ = [
+    "BookScannerService",
+    "FileParserService",
+    "KnowledgeService",
+    "get_file_parser",
+    "get_knowledge_service",
+    "init_file_parser",
+    "init_knowledge_service",
+]
