+++ backend/services/publishing/__init__.py
@@ -1,0 +1,15 @@
+"""Publishing workflows and platform integrations."""
+
+from .oss_service import OSSService, get_oss_service, init_oss_service
+from .publishers import Publisher
+from .xhs_service import XHSService, get_xhs_service, init_xhs_service
+
+__all__ = [
+    "OSSService",
+    "Publisher",
+    "XHSService",
+    "get_oss_service",
+    "get_xhs_service",
+    "init_oss_service",
+    "init_xhs_service",
+]
