+++ backend/services/publishing/publishers/__init__.py
@@ -1,0 +1,4 @@
+# 多平台发布器模块
+from .publisher import Publisher
+
+__all__ = ['Publisher']
