+++ backend/services/publishing/publishers/configs/xiaohongshu.yaml
